Add the OTRS db user
We need a user to be used by OTRS to read the party view to you as
Customers in OTRS. By this reason we add a user in PostgreSQL `otrs`
with the privilege of read the view `party_otrs_view`.

Also I move the vars related with OTRS to the OTRS vars block.