Grant SELECT on otrs_contract view
Without OTRS can't access the view. This however is adds maintenance
costs as we'll need to update the permissions and provision every time
there's a new view to query.

It is safer and more maintainable to move these views within an otrs
schema where only the otrs has read access.