Do not grant SELECT from provision.yml
This fails in development because there are no such views. The
underlying issue is that we applied this change through a sql script
manually executed in production but we did it this way in staging.

Dani and I agree that we must follow the same approach in all
environments, and in this case we're constrained by production. We don't
dare to touch that server too much.