Manage a certificate out of domain
Use the DNS RFC2136 Certbot plugin to create and renew a certificate out
of the server. Using the DNS credentials this plugin allow certbot to
authenticate as DNS manager and modify the TXT DNS entries to allow to
generate a certificate from out of the original server. We use this
plugin (as Pangea told us) to manage the certificate of
suport.somconnexio.coop.

By noew we add this rol in this repository, but in the future, we
pretend to create a public role with the logic to manage this kind of
certificates.

This role does:

- Install certbot.
- Install dns-rfc2136 certbot plugin.
- Copy the DNS credentials to the server.
- Generate a certificate if it doesn't exist
- Configure the renew credentials.

TODO: Use a post hook of certbot to send an email with the certificates
to the development team.